Finished security backend improvements and authentication updates

+++ services.py
@@ -1,12 +1,11 @@
-from datetime import datetime
-
-from flask import session
+import secrets
+from datetime import datetime, timedelta
 from sqlalchemy import or_
 from werkzeug.security import generate_password_hash, check_password_hash
 
 from auth import generate_token
 from database import SessionLocal
-from models import Plan, Meal, User, FoodNorm, Food
+from models import Plan, Meal, User, FoodNorm, Food, Invite
 
 
 def register_trainer(email, password, first_name, last_name):
@@ -14,13 +13,26 @@ def register_trainer(email, password, first_name, last_name):
 
     try:
         existing_user = session.query(User).filter(User.email == email).first()
+
+        if not email.strip():
+            raise ValueError("Email is required!")
+
+        if len(password) < 6:
+            raise ValueError("Password is to short!")
+
         if existing_user:
             raise ValueError("This email already exists!")
+
+        if not first_name.strip():
+            raise ValueError("First name required!")
+
+        if not last_name.strip():
+            raise ValueError("Last name required!")
 
         hashed_password = generate_password_hash(password)
 
         user = User(
-            email=email,
+            email = email.strip().lower(),
             password_hash=hashed_password,
             role="trainer",
             first_name=first_name,
@@ -39,26 +51,176 @@ def register_trainer(email, password, first_name, last_name):
 
 def login_user(email, password):
 
+    email = email.strip().lower()
+
     session = SessionLocal()
     try:
         user = session.query(User).filter(User.email == email).first()
         if not user:
             raise ValueError("Invalid email or password")
 
+        if not user.is_active:
+            raise ValueError("Account is deactivated")
+
         if not check_password_hash(user.password_hash, password):
             raise ValueError("Invalid email or password")
 
         token = generate_token(user)
 
-        return token
+        return token, user
 
     finally:
         session.close()
+
+def create_invite(email, trainer_id):
+
+    session = SessionLocal()
+
+    email = email.strip().lower()
+
+    try:
+        existing_client = session.query(User).filter(User.email == email).first()
+        if existing_client:
+            raise ValueError("This email already exists!")
+
+        existing_invite = session.query(Invite).filter(Invite.used.is_(False), Invite.email == email).first()
+        if existing_invite:
+            raise ValueError("This invitation has been already sent!")
+
+        invite = Invite(
+            email=email,
+            trainer_id=trainer_id,
+            used=False,
+            expires_at=datetime.utcnow() + timedelta(days=7),
+            token=secrets.token_urlsafe(32)
+        )
+
+        session.add(invite)
+        session.commit()
+        session.refresh(invite)
+
+        return invite.to_dict()
+
+    finally:
+        session.close()
+
+def validate_invite_token(token):
+
+    session = SessionLocal()
+
+    current_time = datetime.utcnow()
+
+    try:
+        invite = session.query(Invite).filter(Invite.token == token, Invite.used.is_(False), Invite.revoked.is_(False)).first()
+        if not invite:
+            raise ValueError("Invalid token!")
+
+        invite_expired = current_time > invite.expires_at
+        if invite_expired:
+            raise ValueError("Token expired!")
+
+        return invite.to_dict()
+
+    finally:
+        session.close()
+
+def accept_invite(token, first_name, last_name, password):
+
+    session = SessionLocal()
+
+    try:
+        found_token = session.query(Invite).filter(Invite.token == token).first()
+        if not found_token:
+            raise ValueError("Invalid token!")
+
+        if len(password) < 6:
+            raise ValueError("Password is too short!")
+
+        if not first_name.strip():
+            raise ValueError("First name required")
+
+        if not last_name.strip():
+            raise ValueError("Last name required")
+
+        validate_invite_token(token)
+
+        hashed_password = generate_password_hash(password)
+
+        client = User(email=found_token.email,
+                    password_hash=hashed_password,
+                    first_name=first_name,
+                    last_name=last_name,
+                    role="client",
+                    trainer_id=found_token.trainer_id)
+
+        found_token.used = True
+
+        session.add(client)
+        session.commit()
+        session.refresh(client)
+
+        return client.to_dict()
+
+    finally:
+        session.close()
+
+def revoke_invitation(invite_id, trainer_id):
+
+    session = SessionLocal()
+
+    try:
+        invite = session.query(Invite).filter(Invite.id == invite_id, Invite.trainer_id == trainer_id).first()
+
+        if not invite:
+            raise ValueError("Invitation not found!")
+
+        invite.revoked = True
+
+        session.commit()
+        session.refresh(invite)
+
+        return invite
+
+    finally:
+        session.close()
+
+def get_invitation(trainer_id):
+
+    session = SessionLocal()
+
+    try:
+        invitations = session.query(Invite).filter(Invite.trainer_id == trainer_id, Invite.used == False, Invite.revoked == False).all()
+
+        return [invite.to_dict() for invite in invitations]
+
+    finally:
+        session.close()
+
+
+
+
+def get_client_by_id(client_id, trainer_id):
+
+    session = SessionLocal()
+
+    try:
+        client = (session.query(User).filter(User.id == client_id).filter(User.trainer_id == trainer_id).first())
+        if not client:
+            raise ValueError("Client not found")
+
+        return client.to_dict()
+
+    finally:
+        session.close()
+
 
 def create_plan(name, plan_type, client_id, trainer_id, start_date=None):
     session = SessionLocal()
 
     try:
+        if not name.strip():
+            raise ValueError("Plan name required")
+
         if plan_type not in ["calendar", "template"]:
             raise ValueError("Invalid plan type")
 
@@ -113,7 +275,7 @@ def get_plans(user):
     finally:
         session.close()
 
-def get_plan_by_id(plan_id):
+def get_plan_by_id(plan_id, user):
     session = SessionLocal()
 
     try:
@@ -122,13 +284,19 @@ def get_plan_by_id(plan_id):
         if not plan:
             return None
 
+        if user.role == "trainer" and plan.trainer_id != user.id:
+            return None
+
+        if user.role == "client" and plan.client_id != user.id:
+            return None
+
         return plan.to_dict()
 
     finally:
         session.close()
 
 
-def delete_plan(plan_id):
+def delete_plan(plan_id, user):
 
     session = SessionLocal()
 
@@ -137,6 +305,13 @@ def delete_plan(plan_id):
 
         if not plan:
             raise ValueError("Plan not found")
+
+        if user.role != "trainer":
+            return None
+
+        if user.role == "trainer" and plan.trainer_id != user.id:
+            return None
+
         session.delete(plan)
         session.commit()
 
@@ -145,7 +320,7 @@ def delete_plan(plan_id):
     finally:
         session.close()
 
-def update_plan(plan_id, name, plan_type, client_id, start_date=None):
+def update_plan(plan_id, name, plan_type, client_id, user, start_date=None):
 
     session = SessionLocal()
 
@@ -154,12 +329,33 @@ def update_plan(plan_id, name, plan_type, client_id, start_date=None):
 
         if not plan:
             raise ValueError("Plan not found")
+
+        if not name.strip():
+            raise ValueError("Plan name required")
+
+        if user.role != "trainer":
+            return None
+
+        if user.role == "trainer" and plan.trainer_id != user.id:
+            return None
+
+        client = session.query(User).filter(User.id == client_id).first()
+
+        if not client:
+            raise ValueError("Client not found")
+
+        if client.trainer_id != user.id:
+            raise ValueError("Unauthorized client")
+
         if plan_type not in ["calendar", "template"]:
             raise ValueError("Invalid plan type")
+
         if plan_type == "calendar" and not start_date:
             raise ValueError("Calendar plan type requires start_date")
+
         if plan_type == "template" and start_date:
             raise ValueError("Template plan cannot have start_date")
+
 
         if start_date:
             start_date = datetime.fromisoformat(start_date).date()
@@ -180,13 +376,24 @@ def update_plan(plan_id, name, plan_type, client_id, start_date=None):
         session.close()
 
 
-
-
-def create_meal(name, plan_id, day_number):
+def create_meal(name, plan_id, day_number, user):
 
     session = SessionLocal()
 
     try:
+        plan = session.query(Plan).filter(Plan.id == plan_id).first()
+
+        if not plan:
+            raise ValueError("Plan not found")
+
+        if user.role != "trainer":
+            return None
+
+        if user.role == "trainer" and plan.trainer_id != user.id:
+            return None
+
+        if not name.strip():
+            raise ValueError("Meal name required")
 
         if day_number <= 0:
             raise ValueError("Day can not be zero or negative")
@@ -205,7 +412,7 @@ def create_meal(name, plan_id, day_number):
     finally:
         session.close()
 
-def get_meals_for_plan(plan_id):
+def get_meals_for_plan(plan_id, user):
 
     session = SessionLocal()
 
@@ -215,6 +422,12 @@ def get_meals_for_plan(plan_id):
         if not plan:
             return []
 
+        if user.role == "trainer" and plan.trainer_id != user.id:
+            return None
+
+        if user.role == "client" and plan.client_id != user.id:
+            return None
+
         meals = [m.to_dict() for m in plan.meals]
 
         return meals
@@ -222,7 +435,7 @@ def get_meals_for_plan(plan_id):
     finally:
         session.close()
 
-def delete_meal(meal_id):
+def delete_meal(meal_id, user):
 
     session = SessionLocal()
 
@@ -231,6 +444,12 @@ def delete_meal(meal_id):
 
         if not meal:
             raise ValueError("Meal not found")
+
+        if user.role != "trainer":
+            return None
+
+        if meal.plan.trainer_id != user.id:
+            return None
 
         meal_data = meal.to_dict()
 
@@ -247,30 +466,49 @@ def get_clients_by_trainer(trainer_id):
     session = SessionLocal()
 
     try:
-
-        clients = (
-            session.query(User)
-            .filter(User.trainer_id == trainer_id)
-            .filter(User.role == "client")
-            .all()
-        )
-
+        clients = (session.query(User).filter(User.trainer_id == trainer_id).filter(User.role == "client").all())
         return [c.to_dict() for c in clients]
 
     finally:
         session.close()
+
+def deactivate_client(client_id, trainer_id):
+
+    session = SessionLocal()
+
+    try:
+        client = session.query(User).filter(User.id==client_id, User.trainer_id==trainer_id).first()
+
+        if not client:
+            raise ValueError("Client not found")
+
+        client.is_active = False
+        session.commit()
+        session.refresh(client)
+
+        return client.to_dict()
+
+    finally:
+        session.close()
+
 
 def create_food_norm(user, name, calories, protein, carbs, fat):
 
     session = SessionLocal()
 
     try:
+        if user.role != "trainer":
+            return None
+
         name = name.strip().lower()
 
-        existing = session.query(FoodNorm).filter(FoodNorm.name == name).first()
+        existing = session.query(FoodNorm).filter(FoodNorm.name == name, FoodNorm.created_by == user.id).first()
 
         if existing:
             raise ValueError("Food already exists.")
+
+        if any(float(value or 0) < 0 for value in [calories, protein, fat, carbs]):
+            raise ValueError("Values cannot be negative")
 
         food = FoodNorm(
             name=name,
@@ -306,7 +544,7 @@ def get_food_norms(user):
     finally:
         session.close()
 
-def create_food(meal_id, food_norm_id, grams):
+def create_food(meal_id, food_norm_id, grams, user):
 
     session = SessionLocal()
 
@@ -314,6 +552,12 @@ def create_food(meal_id, food_norm_id, grams):
         meal = session.query(Meal).filter(Meal.id == meal_id).first()
         if not meal:
             raise ValueError("Meal not found")
+
+        if user.role != "trainer":
+            return None
+
+        if meal.plan.trainer_id != user.id:
+            return None
 
         food_norm = session.query(FoodNorm).filter(FoodNorm.id == food_norm_id).first()
         if not food_norm:
@@ -333,7 +577,7 @@ def create_food(meal_id, food_norm_id, grams):
     finally:
         session.close()
 
-def delete_food(food_id):
+def delete_food(food_id, user):
 
     session = SessionLocal()
 
@@ -342,15 +586,23 @@ def delete_food(food_id):
 
         if not food:
             raise ValueError("Food not found")
+
+        if user.role != "trainer":
+            return None
+
+        if food.meal.plan.trainer_id != user.id:
+            return None
+
+        meal = food.meal
 
         session.delete(food)
         session.commit()
 
-        return True
+        return meal.to_dict()
     finally:
         session.close()
 
-def update_food(food_id, grams):
+def update_food(food_id, grams, user):
     session = SessionLocal()
 
     try:
@@ -359,6 +611,13 @@ def update_food(food_id, grams):
         if not food:
             raise ValueError("Food not found")
 
+        if user.role != "trainer":
+            return None
+
+        if food.meal.plan.trainer_id != user.id:
+            return None
+
+        grams = float(grams)
         if grams <= 0:
             raise ValueError("Grams must be positive")
 
@@ -367,7 +626,7 @@ def update_food(food_id, grams):
         session.commit()
         session.refresh(food)
 
-        result = food.to_dict()
+        result = food.meal.to_dict()
         return result
 
     finally:
